+++ datafiles/__init__.py
@@ -4,5 +4,5 @@ from dataclasses import field
 
 from . import converters
 from .decorators import auto, datafile
-from .managers import Missing
-from .models import Model
+from .manager import Missing
+from .model import Model